move logic into MigrationTokenPane folder
