+++ public/app/features/admin/migrate-to-cloud/cloud/MigrationTokenPane/MigrationTokenModal.tsx
