+++ public/app/features/admin/migrate-to-cloud/cloud/MigrationTokenPane/MigrationTokenPane.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { Box, Button, ModalsController, Text } from '@grafana/ui';
 import { t, Trans } from 'app/core/internationalization';
 
-import { useCreateMigrationTokenMutation, useDeleteMigrationTokenMutation, useHasMigrationTokenQuery } from '../api';
-import { InfoItem } from '../shared/InfoItem';
+import { useCreateMigrationTokenMutation, useDeleteMigrationTokenMutation, useHasMigrationTokenQuery } from '../../api';
+import { InfoItem } from '../../shared/InfoItem';
 
 import { DeleteMigrationTokenModal } from './DeleteMigrationTokenModal';
 import { MigrationTokenModal } from './MigrationTokenModal';
