+++ public/locales/en-US/grafana.json
@@ -736,7 +736,7 @@
       "modal-copy-and-close": "Copy to clipboard and close",
       "modal-copy-button": "Copy to clipboard",
       "modal-title": "Migration token created",
-      "status": "Current status: <1></1>",
+      "status": "Current status: <2></2>",
       "title": "Migration token"
     },
     "pdc": {
