+++ public/locales/pseudo-LOCALE/grafana.json
@@ -736,7 +736,7 @@
       "modal-copy-and-close": "Cőpy ŧő čľįpþőäřđ äŉđ čľőşę",
       "modal-copy-button": "Cőpy ŧő čľįpþőäřđ",
       "modal-title": "Mįģřäŧįőŉ ŧőĸęŉ čřęäŧęđ",
-      "status": "Cūřřęŉŧ şŧäŧūş: <1></1>",
+      "status": "Cūřřęŉŧ şŧäŧūş: <2></2>",
       "title": "Mįģřäŧįőŉ ŧőĸęŉ"
     },
     "pdc": {
